Refactored UsagePattternTest class
Broke down the UsagePatternTest class and moved the high number of test cases there into multiple smaller test classes with more organization
